extract unraid-connect concern
