--- packages/unraid-api-plugin-connect/src/model/config.demo.ts
@@ -1,6 +1,0 @@
-import { Field } from '@nestjs/graphql';
-
-export class ConnectDemoConfig {
-    @Field(() => String)
-    demo!: string;
-}
