+++ packages/unraid-api-plugin-connect/src/unraid-connect/connect-settings.service.ts
@@ -20,11 +20,11 @@ import type {
     EnableDynamicRemoteAccessInput,
     RemoteAccess,
     SetupRemoteAccessInput,
-} from '../model/connect.model.js';
+} from './connect.model.js';
 import { ConnectApiKeyService } from '../authn/connect-api-key.service.js';
 import { ConfigType, MyServersConfig } from '../config/connect.config.js';
 import { EVENTS } from '../helper/nest-tokens.js';
-import { DynamicRemoteAccessType, WAN_ACCESS_TYPE, WAN_FORWARD_TYPE } from '../model/connect.model.js';
+import { DynamicRemoteAccessType, WAN_ACCESS_TYPE, WAN_FORWARD_TYPE } from './connect.model.js';
 import { DynamicRemoteAccessService } from '../remote-access/dynamic-remote-access.service.js';
 import { NetworkService } from '../network/network.service.js';
