+++ packages/unraid-api-plugin-connect/src/unraid-connect/connect.module.ts
@@ -6,9 +6,9 @@ import { UserSettingsModule } from '@unraid/shared/services/user-settings.js';
 import { ConnectApiKeyService } from '../authn/connect-api-key.service.js';
 import { ConnectLoginHandler } from '../authn/connect-login.events.js';
 import { ConnectConfigService } from '../config/connect.config.service.js';
-import { ConnectSettingsResolver } from '../resolver/connect-settings.resolver.js';
-import { ConnectResolver } from '../resolver/connect.resolver.js';
-import { ConnectSettingsService } from '../service/connect-settings.service.js';
+import { ConnectSettingsResolver } from './connect-settings.resolver.js';
+import { ConnectResolver } from './connect.resolver.js';
+import { ConnectSettingsService } from './connect-settings.service.js';
 import { RemoteAccessModule } from '../remote-access/remote-access.module.js';
 
 @Module({
